+++ pkgs/objective_c/src/objective_c.c
@@ -22,3 +22,28 @@ bool isValidBlock(ObjCBlockImpl* block) {
          isa == &_NSConcreteAutoBlock || isa == &_NSConcreteFinalizingBlock ||
          isa == &_NSConcreteGlobalBlock || isa == &_NSConcreteWeakBlockVariable;
 }
+
+void finalizeObject(void* isolate_callback_data, void* peer) {
+  // objc_release works for Objects and Blocks.
+  objc_release(peer);
+}
+
+Dart_FinalizableHandle newFinalizableHandle(Dart_Handle owner,
+                                                  ObjCObject* object) {
+  return Dart_NewFinalizableHandle_DL(owner, object, 0, finalizeObject);
+}
+
+void deleteFinalizableHandle(Dart_FinalizableHandle handle, Dart_Handle owner) {
+  Dart_DeleteFinalizableHandle_DL(handle, owner);
+}
+
+void finalizeMalloc(void* isolate_callback_data, void* peer) {
+  free(peer);
+}
+
+bool* newFinalizableBool(Dart_Handle owner) {
+  bool* pointer = (bool*)malloc(1);
+  *pointer = false;
+  Dart_NewFinalizableHandle_DL(owner, pointer, 1, finalizeMalloc);
+  return pointer;
+}